+++ src/frontend/src/features/settings/index.ts
@@ -1,1 +1,2 @@
 export { SettingsButton } from './components/SettingsButton'
+export { SettingsDialog } from './components/SettingsDialog'
