--- src/frontend/src/primitives/AlertDialog.tsx
@@ -1,5 +1,0 @@
-import { Dialog, type DialogProps } from './Dialog'
-
-export const AlertDialog = (props: DialogProps) => {
-  return <Dialog role="alertdialog" {...props} />
-}
