+++ src/frontend/src/primitives/Dialog.tsx
@@ -1,16 +1,15 @@
-import { useContext, type ReactNode } from 'react'
+import { styled } from '@/styled-system/jsx'
+import { RiCloseLine } from '@remixicon/react'
+import { t } from 'i18next'
 import {
   Dialog as RACDialog,
-  DialogTrigger,
-  Modal,
-  type DialogProps as RACDialogProps,
   ModalOverlay,
-  OverlayTriggerStateContext,
+  Modal,
+  type DialogProps,
+  Heading,
 } from 'react-aria-components'
-import { RiCloseLine } from '@remixicon/react'
-import { useTranslation } from 'react-i18next'
-import { styled } from '@/styled-system/jsx'
-import { Button, Box, Div, VerticallyOffCenter } from '@/primitives'
+import { Div, Button, Box, VerticallyOffCenter } from '@/primitives'
+import { text } from './Text'
 
 const StyledModalOverlay = styled(ModalOverlay, {
   base: {
@@ -47,71 +46,56 @@ const StyledRACDialog = styled(RACDialog, {
     pointerEvents: 'none',
   },
 })
-
-export type DialogProps = {
-  children: [
-    trigger: ReactNode,
-    dialogContent:
-      | (({ close }: { close: () => void }) => ReactNode)
-      | ReactNode,
-  ]
-} & RACDialogProps
-
-/**
- * a Dialog is a tuple of a trigger component (most usually a Button) that toggles some interactive content in a Dialog on top of the app. You should mostly use a DialogContent as second child.
- */
-export const Dialog = ({ children, ...dialogProps }: DialogProps) => {
-  const { t } = useTranslation()
+export const Dialog = ({
+  title,
+  children,
+  ...dialogProps
+}: DialogProps & { title: string }) => {
   const isAlert = dialogProps['role'] === 'alertdialog'
-  const [trigger, dialogContent] = children
   return (
-    <DialogTrigger>
-      {trigger}
-      <StyledModalOverlay
-        isKeyboardDismissDisabled={isAlert}
-        isDismissable={!isAlert}
-      >
-        <StyledModal>
-          <StyledRACDialog {...dialogProps}>
-            {({ close }) => (
-              <VerticallyOffCenter>
-                <Div
-                  width="fit-content"
-                  maxWidth="full"
-                  margin="auto"
-                  pointerEvents="auto"
-                >
-                  <Box size="sm" type="dialog">
-                    {typeof dialogContent === 'function'
-                      ? dialogContent({ close })
-                      : dialogContent}
-                    {!isAlert && (
-                      <Div position="absolute" top="0" right="0">
-                        <Button
-                          invisible
-                          size="xs"
-                          onPress={() => close()}
-                          aria-label={t('closeDialog')}
-                        >
-                          <RiCloseLine />
-                        </Button>
-                      </Div>
-                    )}
-                  </Box>
-                </Div>
-              </VerticallyOffCenter>
-            )}
-          </StyledRACDialog>
-        </StyledModal>
-      </StyledModalOverlay>
-    </DialogTrigger>
+    <StyledModalOverlay
+      isKeyboardDismissDisabled={isAlert}
+      isDismissable={!isAlert}
+    >
+      <StyledModal>
+        <StyledRACDialog {...dialogProps}>
+          {({ close }) => (
+            <VerticallyOffCenter>
+              <Div
+                width="fit-content"
+                maxWidth="full"
+                margin="auto"
+                pointerEvents="auto"
+              >
+                <Box size="sm" type="dialog">
+                  <Heading
+                    slot="title"
+                    level={1}
+                    className={text({ variant: 'h1' })}
+                  >
+                    {title}
+                  </Heading>
+                  {typeof children === 'function'
+                    ? children({ close })
+                    : children}
+                  {!isAlert && (
+                    <Div position="absolute" top="0" right="0">
+                      <Button
+                        invisible
+                        size="xs"
+                        onPress={() => close()}
+                        aria-label={t('closeDialog')}
+                      >
+                        <RiCloseLine />
+                      </Button>
+                    </Div>
+                  )}
+                </Box>
+              </Div>
+            </VerticallyOffCenter>
+          )}
+        </StyledRACDialog>
+      </StyledModal>
+    </StyledModalOverlay>
   )
-}
-
-export const useCloseDialog = () => {
-  const dialogState = useContext(OverlayTriggerStateContext)
-  if (dialogState) {
-    return dialogState.close
-  }
-  return null
 }
