+++ src/frontend/src/primitives/index.ts
@@ -1,11 +1,10 @@
 export { A } from './A'
-export { AlertDialog } from './AlertDialog'
 export { Badge } from './Badge'
 export { Bold } from './Bold'
 export { Box } from './Box'
 export { Button } from './Button'
-export { Dialog, useCloseDialog } from './Dialog'
-export { DialogContent } from './DialogContent'
+export { useCloseDialog } from './useCloseDialog'
+export { Dialog } from './Dialog'
 export { Div } from './Div'
 export { Field } from './Field'
 export { Form } from './Form'
